Split flag parsing between connector and PowerShell arguments
Prior to this any "-h" or "--help" got parsed from any additional PowerShell argument.

We can not disable parsing for those flags as they are hardcoded in.

Now the list of arguments is split by the first shorthand flag, so flags like "-C" or "-Command" will split and further arguments are then parsed as for powershell.exe.